refactor(menu): mv consts/index to types
