+++ packages/components/menu/menu-item.tsx
@@ -1,6 +1,6 @@
 import { defineComponent, computed, inject, onMounted, ref, toRefs, getCurrentInstance } from 'vue';
 import props from './menu-item-props';
-import { TdMenuInterface, TdSubMenuInterface } from './consts';
+import { TdMenuInterface, TdSubMenuInterface } from './types';
 import { useTNodeJSX, useContent } from '../hooks/tnode';
 import useRipple from '../hooks/useRipple';
 import { usePrefixClass } from '../hooks/useConfig';
